requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
jinja2==3.1.2
python-multipart==0.0.6
openai>=1.3.8
python-dotenv==1.0.1
httpx>=0.27.0 